feat: Add "Hackerland Radio Transmitters" first solution to the repo
Note: Takes a while due to the list comprehension, but it works. Times
out on two (2) cases, can try to fix with the helper function later,
but was not working as of today. I did correctly identify the type of
algorithm needed (greedy) and the "safe" step, so that worked out well!