feat: Add basic solution to "Project Euler 235" from HackerRank
Note: Does not work on their site, since they do not allow you to
import scipy.optimize. Pretty sure you need to find a numerical
solution for equation, and also sure that the given equation is
correct. Could eventually copy scipy's "fsolve" code to get it to work,
but there might be a trick...