feat: Add solution to "Maximum Index Product" problem from HackerRank
Note: Good to see the problems where it timed out to get fixes - had
to short-circuit the searches before and after by keeping track of
the largest item in each part of the array. Final trick was to check
for the largest item remaining when the current value is greater than
all previous values - otherwise it would skip the calculation of the
largest value remaining to the right.